Task1/requirements.txt
absl-py==0.6.1
alabaster==0.7.12
astor==0.7.1
astroid==2.1.0
Babel==2.6.0
backcall==0.1.0
bleach==3.0.2
certifi==2018.10.15
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29
dask==0.20.2
decorator==4.3.0
defusedxml==0.5.0
docutils==0.14
entrypoints==0.2.3
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
idna==2.7
imagesize==1.1.0
imgaug==0.2.6
ipykernel==5.1.0
ipyparallel==6.2.3
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keras-resnet==0.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nose==1.3.7
notebook==5.7.2
numpy==1.15.4
opencv-python==3.4.3.18
packaging==18.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
progressbar==2.5
progressbar2==3.38.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pylint==2.2.2
pyparsing==2.3.0
python-dateutil==2.7.5
python-utils==2.3.0
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.20.1
scikit-image==0.14.1
scipy==1.1.0
Send2Trash==1.5.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
tensorboard==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
